Add developmental/ecological tooling, spatial food, and interactive app
Simulation & ecology:
- Spatial food growth: `growth_rate` accepts an expression of x,y (per-cell
  `FoodType.growth_field`); rate-1 kernel + pointwise field, back-compatible with scalars
- Spatially-varying + named chemical signatures (`ft-*`/`agents.chemical_signature`)
- Satiation via `agents.eat_energy_fraction`; distance/speed tracking; age-spread init
- Fixes: FFT growth noise floor, corpse chemical emission, AgentConfig dropped fields,
  body-size energy exploit, immobile-agent / reproduction bugs

Devo & motor:
- `ciliated_torque` motor interface (holonomic, rigid-body torque, rounded corners)
- Cilium size from motor gene `m` (continuous, threshold removed) in ciliated interfaces

Evo / QD:
- MAP-Elites over RAND (`src/evo/qd.py`, `qd_rand.py`): swappable fitness/descriptors,
  taxis / MNIST / plasticity / embodiment scoring, `develop_with_trace`

Environments & viz:
- MiniTaxis / MiniMultiTaxis single-agent arenas; retina sensory interface
- Rollout + development animations, network / repertoire / dev-trajectory plots
- Interactive Panel web app (build/run/watch, overlays, brush, inspector, spawn, speed)
- Checkpointing; POSITION_DTYPE (float32 positions); README + CLAUDE.md config reference

+++ src/evo/core.py
@@ -19,7 +19,8 @@ class MutationModel(eqx.Module):
 		return Genotype(neural_params, size, genotype.chemical_emission_signature) # chemical signature is left unchanged
 	#-------------------------------------------------------------------
 	def mutate_size(self, size: Float, key: jax.Array)->Float:
-		return size + jr.normal(key)*self.sigma_size
+		# jr.normal is float32; keep the mutation in the genotype's dtype (body size is float16)
+		return (size + jr.normal(key)*self.sigma_size).astype(size.dtype)
 	#-------------------------------------------------------------------
 	def mutate_neural_params(self, params: NeuralParams, key: jax.Array)->NeuralParams:
 		raise NotImplementedError
